# Increase timeout in benchmark case