Fix #1436

This pull request adjusts timeout durations in benchmark test cases to improve reliability under high workloads. The changes increase the timeout values for creating and deleting nodes and pods in both standard and "hack" benchmark scenarios.

### Timeout adjustments for benchmark tests:

* [`test/e2e/benchmark.go`](diffhunk://#diff-9433aca2d6a2499b55511829069fe44703d8f16d7bf479f0c1637fbdc7107fb3L220-R220): Increased the timeout for deleting pods in the `CaseBenchmark` function from 60 seconds to 90 seconds to handle larger workloads.

* `test/e2e/benchmark_hack.go`: 
  - Increased the timeout for deleting pods in the `CaseBenchmarkWithHack` function from 20 seconds to 30 seconds.
Extend the timeout for pod deletion to ensure successful completion during benchmarks.